Day 4/Task04/requirements.txt
pypdf
streamlit
python-dotenv
openai-agents
openai
pydantic